Login/signup: make sign up behave like log in.

+++ server/controllers/AuthController.js
@@ -1,6 +1,6 @@
 const services = require('../services');
-const SIGN_UP = 'signup';
-const LOGIN = 'login';
+const { GithubService, UsersService } = services;
+const OAUTH = 'oauth';
 
 const capitalizeServiceString = service =>
     `${service.charAt(0).toUpperCase()}${service.slice(1)}`;
@@ -16,64 +16,69 @@ function destroySession(req, res) {
     });
 }
 
+function resolveService(service) {
+    return (
+        services[`${capitalizeServiceString(service)}Service`] || GithubService
+    );
+}
+
 module.exports = {
     async oauth(req, res) {
-        const { service, referer, type } = req.query;
-        const authService =
-            services[`${capitalizeServiceString(service)}Service`] ||
-            services.GithubService;
+        const { service, referer } = req.query;
+        const authService = resolveService(service);
         req.session.referer = referer;
-        req.session.authType = type;
-        const loginURL = authService.url;
-        res.redirect(303, loginURL);
+        req.session.authType = OAUTH;
+        res.redirect(303, authService.url);
         res.end();
     },
 
     async authorize(req, res) {
         const { service, code } = req.query;
-        const authService =
-            services[`${capitalizeServiceString(service)}Service`] ||
-            services.GithubService;
+        const authService = resolveService(service);
 
         req.session.accessToken = await authService.authorize(code);
 
         let userToAuthorize;
+        let authorizationError;
         try {
             userToAuthorize = await authService.getUser({
                 accessToken: req.session.accessToken
             });
         } catch (error) {
-            res.status(401);
-            res.end();
-            return;
+            authorizationError = error;
         }
         let authorized = false;
-        if (req.session.authType === SIGN_UP) {
-            const userSaved = await services.UsersService.signupUser({
-                accessToken: req.session.accessToken,
-                user: userToAuthorize
-            });
-            if (userSaved) {
-                authorized = true;
-                userToAuthorize = userSaved;
-            }
-        } else if (req.session.authType === LOGIN) {
-            const { name: fullname, username, email } = userToAuthorize;
+        if (req.session.authType === OAUTH) {
+            let user;
 
-            const dbUser = await services.UsersService.getUserAndUpdateRoles({
-                accessToken: req.session.accessToken,
-                user: {
+            // If this is a known user that we can authorize...
+            if (userToAuthorize) {
+                const { name: fullname, username, email } = userToAuthorize;
+                user = await UsersService.getUser({
                     fullname,
                     username,
                     email
-                }
-            });
-            if (dbUser) {
+                });
+            }
+
+            // ...otherwise, add them as a new user.
+            if (!user) {
+                user = await UsersService.signupUser({
+                    accessToken: req.session.accessToken,
+                    user: userToAuthorize
+                });
+            }
+
+            // Assuming we now have a user, assign updated roles
+            if (user) {
                 authorized = true;
-                userToAuthorize = dbUser;
+                userToAuthorize = await UsersService.getUserAndUpdateRoles({
+                    accessToken: req.session.accessToken,
+                    user
+                });
             }
         }
-        if (authorized) {
+        if (authorized && userToAuthorize) {
             req.session.user = userToAuthorize;
             res.redirect(303, `${req.session.referer}`);
             res.end(() => {
@@ -81,8 +86,13 @@ module.exports = {
                 delete req.session.authType;
             });
         } else {
-            res.redirect(303, `${req.session.referer}/signupInstructions`);
-            res.end(() => destroySession(req, res));
+            if (authorizationError) {
+                res.status(401);
+                res.end();
+            } else {
+                res.redirect(303, `${req.session.referer}/signupInstructions`);
+                res.end(() => destroySession(req, res));
+            }
         }
     },
 
@@ -96,7 +106,7 @@ module.exports = {
         res.end();
     },
 
-    logout(req, res) {
+    signout(req, res) {
         destroySession(req, res);
     }
 };
